[SofaKerner][Modules] FIX: some warnings of uint/int conversion.

+++ modules/SofaExporter/MeshExporter.cpp
@@ -232,7 +232,7 @@ bool MeshExporter::writeMeshVTKXML()
     //write points
     outfile << "      <Points>\n";
     outfile << "        <DataArray type=\"Float32\" NumberOfComponents=\"3\" format=\"ascii\">\n";
-    for (int i=0 ; i<nbp; i++)
+    for (size_t i=0 ; i<nbp; i++)
     {
         outfile << "          " << pointsPos[i] << "\n";
     }
@@ -391,7 +391,7 @@ bool MeshExporter::writeMeshVTK()
     outfile << "POINTS " << nbp << " float\n";
 
     //Write Points
-    for (int i=0 ; i<nbp; i++)
+    for (size_t i=0 ; i<nbp; i++)
     {
         outfile << pointsPos[i] << "\n";
     }
@@ -504,7 +504,7 @@ bool MeshExporter::writeMeshGmsh()
     outfile << "$Nodes\n";
 
     outfile << nbp << "\n";
-    for (int i=0 ; i<nbp; i++)
+    for (size_t i=0 ; i<nbp; i++)
     {
         outfile << 1+i << ' ' << pointsPos[i] << "\n";
     }
@@ -614,7 +614,7 @@ bool MeshExporter::writeMeshNetgen()
 
     //Write Points
     outfile << nbp << "\n";
-    for (int i=0 ; i<nbp; i++)
+    for (size_t i=0 ; i<nbp; i++)
     {
         outfile << pointsPos[i] << "\n";
     }
@@ -710,7 +710,7 @@ bool MeshExporter::writeMeshTetgen()
     // <# of points> <dimension (must be 3)> <# of attributes> <# of boundary markers (0 or 1)>
     outfile << nbp << ' ' << 3 << ' ' << 0 << ' ' << 0 << "\n";
     // <point #> <x> <y> <z> [attributes] [boundary marker]
-    for (int i=0 ; i<nbp; i++)
+    for (size_t i=0 ; i<nbp; i++)
     {
         outfile << i+1 << ' ' << pointsPos[i] << "\n";
     }
